refactor: group io logic
Currently io logic was mixed in with other stuff in the utils folder. Now that we have a few io related modules it makes sense to group them in a folder.

Moved all io related modules in the an `io` folder.
