+++ src/cmd-add.ts
@@ -6,7 +6,7 @@ import {
   ManifestSaveError,
   tryLoadProjectManifest,
   trySaveProjectManifest,
-} from "./utils/project-manifest-io";
+} from "./io/project-manifest-io";
 import { EnvParseError, parseEnv } from "./utils/env";
 import {
   compareEditorVersion,
